Fix a type conversion bug and the inclusion of sub-derivations in group `all`
The set of default conversions is expanded with a conversion for the Boolean types
and they have been expanded with the actual mapping the rule engine uses for the
Fact.valueType. (The rule-engine uses Scala's weakTypeOf.toString, as opposed to
the getClass.getName). Fixes #6.

The DerivationsService had no way to distinguish between top-level derivations and
those used in sub-derivation constructs. The latter should not be included on the
top level, because they may (and have) attempt to satisfy the same outputs. While
this calls for a better way to namespace facts and derivations, for now the service
differentiates between them based on whether they implement ElementBerekening
or not. Fixes #7.